# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LearnVault API
  version: 1.0.0
  description: Backend API for LearnVault frontend and integrations.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ZodIssue:
      type: object
      required:
        - code
        - path
        - message
      properties:
        code:
          type: string
          example: invalid_type
        path:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          example:
            - walletAddress
        message:
          type: string
          example: Required
    ValidationErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ZodIssue"
...
  responses:
    BadRequestError:
      description: Bad request — validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/scholars/leaderboard:
    get:
      tags:
        - Scholars
      summary: Get scholars leaderboard
      description:
        Returns a paginated ranking of scholars by LRN balance, with optional
        search.
      responses:
        "200":
          description: Leaderboard fetched successfully
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/scholars/{address}:
    get:
      tags:
        - Scholars
      summary: Get scholar profile
      description:
        Returns a scholar's on-chain balances, enrolled courses, milestone
        stats, credentials, and join date.
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scholar profile
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/scholars/{address}/milestones:
    get:
      tags:
        - Scholars
      summary: Get milestones for a scholar
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scholar milestones
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/scholars/{address}/credentials:
    get:
      tags:
        - Scholars
      summary: Get credentials for a scholar
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scholar credentials
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/scholars/{address}/follow:
    post:
      tags:
        - Scholars
      summary: Follow a scholar
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Scholars
      summary: Unfollow a scholar
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags:
        - Scholars
      summary: Get follow status and counts
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/admin/milestones/pending:
    get:
      tags:
        - Admin
      summary: List all unverified milestone reports
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of pending milestone reports
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/admin/milestones/{id}:
    get:
      tags:
        - Admin
      summary: Get milestone report details and evidence
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestone report with audit log
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/admin/milestones/{id}/approve:
    post:
      tags:
        - Admin
      summary: Approve a milestone report and trigger contract call
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestone approved
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Report already processed
  /api/admin/milestones/{id}/reject:
    post:
      tags:
        - Admin
      summary: Reject a milestone report with a reason
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Milestone rejected
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Report already processed
  /api/milestones/submit:
    post:
      tags:
        - Milestones
      summary: Scholar submits a milestone report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scholarAddress
                - course_id
                - milestone_id
              properties:
                scholarAddress:
                  type: string
                course_id:
                  type: string
                milestone_id:
                  type: integer
                evidenceGitHub:
                  type: string
                evidenceIpfsCid:
                  type: string
                evidenceDescription:
                  type: string
      responses:
        "201":
          description: Report submitted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          description: Report already submitted for this milestone
        "429":
          description: Rate limit exceeded
  /api/milestones/resubmit:
    post:
      tags:
        - Milestones
      summary: Scholar resubmits a rejected milestone report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: integer
                evidenceGithub:
                  type: string
                evidenceIpfsCid:
                  type: string
                evidenceDescription:
                  type: string
      responses:
        "200":
          description: Report resubmitted
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          description: Report not found
        "429":
          description: Rate limit exceeded
...